requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
pycodestyle==2.9.1
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
